api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
instagrapi==2.1.2
python-multipart==0.0.20
pydantic==2.7.1
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
databases==0.9.0
sqlalchemy==2.0.36
httpx==0.28.1
python-dotenv==1.0.1
pillow==11.0.0
